+++ lib/app.dart
@@ -5,12 +5,11 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get/get.dart';
-import 'package:mangodevelopment/view/login/signUp.dart';
-import 'package:mangodevelopment/view/refrigerator/addFoodQuick.dart';
+import 'package:mangodevelopment/view/login/addUserInfo.dart';
+import 'package:mangodevelopment/view/login/guide.dart';
 
 import 'color.dart';
 import 'landing.dart';
-import 'view/trade/friend/friendList.dart';
 import 'dart:async';
 
 var deviceWidth = 411.0;
@@ -91,8 +90,7 @@ class _MangoAppState extends State<MangoApp> {
   Widget build(BuildContext context) {
     return ScreenUtilInit(
       builder: () => GetMaterialApp(
-        // home: Landing(),
-        home: AddFoodQuickPage(),
+        home: Landing(),
         theme: _mangoTheme,
         // getPages: [GetPage(name: 'FriendList', page: () => FriendListPage())],
       ),
