+++ lib/landing.dart
@@ -1,6 +1,7 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:mangodevelopment/view/login/addUserInfo.dart';
 import 'package:mangodevelopment/view/widget/mangoIndicator.dart';
 import 'package:mangodevelopment/viewModel/authentication.dart';
 import 'package:shared_preferences/shared_preferences.dart';
@@ -45,10 +46,15 @@ class _LandingState extends State<Landing> {
                 return LogInPage(title: 'hi');
               } else {
                 return FutureBuilder(
-                    future: authController.hasData(authController.user!.uid),
+                    future:
+                        authController.hasAlarmData(authController.user!.uid),
                     builder: (context, snapshot) {
+                      print(snapshot);
+                      print(snapshot.data);
                       if (snapshot.hasData == false) {
                         return MangoCircularIndicator();
+                      } else if (snapshot.data == false) {
+                        return AddUserInfoPage();
                       } else {
                         //_prefs!.setString('id', authController.user!.uid);
                         return HomePage(title: 'hi');
